# ipx/docs/swagger.yaml
basePath: /
definitions:
...
  v4.BuildERC20EIP1559TransactionRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      contract:
        example: 0x5FbDB2315678afecb367f032d93F642f64180aa3
        type: string
      from:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  v4.BuildERC20EIP1559TransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v4.BuildERC20LegacyTransactionRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      contract:
        example: 0x5FbDB2315678afecb367f032d93F642f64180aa3
        type: string
      from:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  v4.BuildERC20LegacyTransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v4.BuildNativeEIP1559TransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v4.BuildNativeEIP1559TransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v4.BuildNativeLegacyTransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v4.BuildNativeLegacyTransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
...
  /evm/v4/transaction/erc20/eip1559:
    post:
      consumes:
      - application/json
      description: Builds transfer(address,uint256) calldata, estimates gas, signs
        with the configured key, broadcasts, and returns unsigned_rlp, signed_rlp,
        tx_hash, r, s, v
      parameters:
      - description: ERC-20 transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v4.BuildERC20EIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BuildERC20EIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build, sign, and send ERC-20 EIP-1559 transfer tx
      tags:
      - transaction
  /evm/v4/transaction/erc20/legacy:
    post:
      consumes:
      - application/json
      description: Builds transfer(address,uint256) calldata, estimates gas, signs
        with the configured key, broadcasts, and returns unsigned_rlp, signed_rlp,
        tx_hash, r, s, v
      parameters:
      - description: ERC-20 transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v4.BuildERC20LegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BuildERC20LegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build, sign, and send ERC-20 legacy transfer tx
      tags:
      - transaction
  /evm/v4/transaction/native/eip1559:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-1559 native transfer,
        broadcasts it, and returns unsigned_rlp, signed_rlp, tx_hash, r, s, v
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v4.BuildNativeEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BuildNativeEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build, sign, and send EIP-1559 transfer tx
      tags:
      - transaction
  /evm/v4/transaction/native/legacy:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-155 legacy native
        transfer, broadcasts it, and returns unsigned_rlp, signed_rlp, tx_hash, r,
        s, v
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v4.BuildNativeLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BuildNativeLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build, sign, and send legacy transfer tx
      tags:
      - transaction
swagger: "2.0"